chore(react): chore: subPath 적용
