+++ .changeset/fluffy-ads-train.md
@@ -1,0 +1,5 @@
+---
+'@modern-kit/react': major
+---
+
+chore(react): chore: subPath 적용 - @ssi02014
