+++ packages/react/build.utils.mjs
@@ -1,0 +1,32 @@
+const getEntryFileNames = (name, format, path) => {
+  const cleanName = name.replace(`${path}-`, '');
+  return `${path}/${cleanName}/${cleanName}.${format}`;
+};
+
+export const getFormatEntryFileNames = (chunkInfo, format) => {
+  const prefixMap = {
+    _internal: '_internal',
+    utils: 'utils',
+    components: 'components',
+    hooks: 'hooks',
+  };
+
+  for (const [key, value] of Object.entries(prefixMap)) {
+    if (chunkInfo.name === key) {
+      return `${key}/index.${format}`;
+    }
+
+    if (chunkInfo.name.startsWith(value)) {
+      return getEntryFileNames(chunkInfo.name, format, key);
+    }
+  }
+
+  return `[name].${format}`; // 기본 파일 이름
+};
+
+export const getSubEntryMap = (keys, path) => {
+  return keys.reduce((acc, entry) => {
+    acc[`${path}-${entry}`] = `./src/${path}/${entry}`;
+    return acc;
+  }, {});
+};
