+++ packages/react/src/components/InfiniteScroll/InfiniteScroll.spec.tsx
