--- packages/react/src/hooks/useTimeout/useTimeout.types.ts
@@ -1,10 +1,0 @@
-export interface TimeoutOptions {
-  delay: number;
-  enabled?: boolean;
-}
-
-export interface UseTimeoutReturnType {
-  set: () => void;
-  clear: () => void;
-  reset: () => void;
-}
