+++ packages/react/subEntries.mjs
@@ -1,0 +1,78 @@
+export const componentsSubEntryKeys = [
+  'AspectRatio',
+  'ClientGate',
+  'Mounted',
+  'DebounceHandler',
+  'FallbackLazyImage',
+  'IfElse',
+  'InfiniteScroll',
+  'InView',
+  'Iterator',
+  'LazyImage',
+  'OutsidePointerDownHandler',
+  'Portal',
+  'SeparatedIterator',
+  'Slot',
+  'When',
+  'SwitchCase',
+  // 컴포넌트 추가 시 추가
+];
+
+export const hooksSubEntryKeys = [
+  'useAsyncEffect',
+  'useAsyncProcessQueue',
+  'useBeforeUnload',
+  'useBlockMultipleAsyncCalls',
+  'useClipboard',
+  'useColorScheme',
+  'useCounter',
+  'useCycleList',
+  'useDebounce',
+  'useDebouncedInputValue',
+  'useDebouncedState',
+  'useDocumentTitle',
+  'useEventListener',
+  'useFileReader',
+  'useFocus',
+  'useForceUpdate',
+  'useHover',
+  'useImageStatus',
+  'useIntersectionObserver',
+  'useInterval',
+  'useIsMounted',
+  'useIsomorphicLayoutEffect',
+  'useKeyDown',
+  'useLocalStorage',
+  'useMediaQuery',
+  'useMergeRefs',
+  'useMouse',
+  'useNetwork',
+  'useOutsidePointerDown',
+  'usePreferredColorScheme',
+  'usePreservedCallback',
+  'usePreservedState',
+  'usePrevious',
+  'useResizeObserver',
+  'useScrollLock',
+  'useScrollTo',
+  'useSessionStorage',
+  'useStep',
+  'useStepState',
+  'useThrottle',
+  'useTimeout',
+  'useToggle',
+  'useToggleState',
+  'useUnmount',
+  'useUserAgent',
+  'useVhProperty',
+  'useVisibilityChange',
+  'useWindowSize',
+  'useDependencyTimeout',
+  // 신규 훅 추가 시 추가
+];
+
+export const utilsEntryKey = [
+  'mergeRefs',
+  'polymorphicForwardRef',
+  // 신규 유틸 모듈 추가 시 추가
+];
